builder-pattern-service/hello_world/requirements.txt
fastapi==0.104.1
mangum>=0.17.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.2
langchain-anthropic>=0.1.1
python-dotenv>=1.0.0
openai>=1.6.1
anthropic>=0.7.7
pydantic>=2.5.2
boto3>=1.34.11